# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "One Medy API Documentation",
        "description": "One Medy APi Service to be consumed by client Application",
        "contact": {
            "email": "info@walexbiznig.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/labtestrequest/add": {
            "post": {
                "tags": [
                    "Labtest Request"
                ],
                "summary": "Create a new test request to the Labs.",
                "operationId": "cfc80e17f5ffaaf3b9a6a7e0d1c78201",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateLabRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/labtestrequests": {
            "get": {
                "tags": [
                    "Labtest Request"
                ],
                "summary": "Return a json array of all the tests requests sent to the labs.",
                "description": "Returns all test requests both the completed and uncompleted sent to the Laboratory. If iscompleted is placed in the URL as parameter with a value of\r\n     * 1, it will return all completed tests. The opposite happens if the value of iscompleted is made 0",
                "operationId": "getAllTestRequests",
                "parameters": [
                    {
                        "name": "iscompleted",
                        "in": "parameter",
                        "description": "Status of the Lab test",
                        "required": false
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/doctor/labtest/complete/{id}": {
            "patch": {
                "tags": [
                    "Labtest Request"
                ],
                "summary": "Set a lab test request as completed.",
                "operationId": "b526f285aee456f3e59378bcbfb92393",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateLabRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/labtestrequest/tenant": {
            "get": {
                "tags": [
                    "Labtest Request"
                ],
                "summary": "Get list of labtest request by tenant id",
                "description": "Returns list all labtest requests by passing a tenant id query parameter",
                "operationId": "TenantId",
                "parameters": [
                    {
                        "name": "tenant_id",
                        "in": "query",
                        "description": "Tenant id"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/labtestrequest/result/{labtest_id}": {
            "patch": {
                "tags": [
                    "Labtest Request"
                ],
                "summary": "Post labtest result and conclusion",
                "description": "Post labtest result and conclusion",
                "operationId": "TenantId",
                "parameters": [
                    {
                        "name": "tenant_id",
                        "in": "query",
                        "description": "Tenant id"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/notification/add": {
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "system admin creates notification",
                "operationId": "1b54ddea847a9e143db01f32a5c0621f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateNotificationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/notification/get/{id}": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get notification by user id",
                "description": "Returns notification by id",
                "operationId": "getNotificationById",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Size"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/v1/pharmacyorder/tenant": {
            "get": {
                "tags": [
                    "Pharmacy Orders"
                ],
                "summary": "Get list of pharmacy orders by tenant id",
                "description": "Returns list all pharmacy orders by tenant id",
                "operationId": "TenantId",
                "parameters": [
                    {
                        "name": "tenant_id",
                        "in": "query",
                        "description": "Tenant Id"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateLabRequest": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "labtest_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "tenant_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "delivery_type": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "integer": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/CreateLabRequest"
                    }
                ]
            },
...
            "CreateNotificationRequest": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "user_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "title": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "type": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "description": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "isread": {
                                "type": "boolean",
                                "format": "boolean"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/CreateNotificationRequest"
                    }
                ]
            },
...
        },
...
    },
...
    "security": [
        []
    ]
}